// example-app/src/swagger-generated/accounts/api.ts
export interface AccountRequest {
    /**
     * Specified date and time when authorization expires. This can not be combined with transactionFrom and transactionTo. Expiration date time can not be more than 90 days in to the future. All dates in the JSON payloads are represented in ISO 8601 date-time format. All date-time fields in responses must include the timezone.
     * @type {Date}
     * @memberof AccountRequest
     */
    expires?: Date;
    /**
     * Specified start date for transaction query period. The date cannot exceed 730 days into past.
     * @type {string}
     * @memberof AccountRequest
     */
    transactionFrom?: string;
    /**
     * Specified end date and time for transaction query period. The data must not exceed current date. Date is inclusive.
     * @type {string}
     * @memberof AccountRequest
     */
    transactionTo?: string;
}
// ...
export interface Authorization {
    /**
     * 
     * @type {string}
     * @memberof Authorization
     */
    authorizationId: string;
    /**
     * 
     * @type {Date}
     * @memberof Authorization
     */
    created: Date;
    /**
     * 
     * @type {Date}
     * @memberof Authorization
     */
    modified?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Authorization
     */
    authorized?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Authorization
     */
    expires?: Date;
    /**
     * 
     * @type {string}
     * @memberof Authorization
     */
    transactionFrom?: string;
    /**
     * 
     * @type {string}
     * @memberof Authorization
     */
    transactionTo?: string;
    /**
     * 
     * @type {AuthorizationStatus}
     * @memberof Authorization
     */
    status: AuthorizationStatus;
}

/**
 * 
 * @export
 * @enum {string}
 */
export enum AuthorizationStatus {
    Unauthorized = 'Unauthorized',
    Authorized = 'Authorized',
    Rejected = 'Rejected',
    Revoked = 'Revoked',
    Expired = 'Expired'
}